Shutter and Beam Parking Updates
Added manual control of stim path shutter to GUI, and color coding of
'Stim' button to reflect shutter status. Added a mirror-prepositioning
feature to the 'arm' callback. This prepositions X + Y mirrors at the
start of a stimulation signal, and is useful for the first stimulation
after selecting a new ROI, since the mirrors will be in the vicinity of
the old ROI at the beginning of the first stimulation.