modified parking behavior
Modified code so mirrors are no longer parked on next cell to be
stimulated. Instead parked at +4V on both channels (should be on edge or
outside of optical path of scope, but ok because of pockels blanking
laser) at end of each trial (but not btw stim within trial). Also
disabled mirror pre-positioning before a trial, and added a 2ms pockels
blanking to the beginning of a trial, to allow the stim mirrors to move
from parked to cell location before beam turns on.