fixed currentStim counter bug
onDmd sequencer uses numSequences to track the current stimulation
number. Fixed a bug where I used that variable to save the framesITI
parameter, forgetting it's use. Instead framesITI is saved in the
argument for the stimSequencer abort callback (this argument is not used
in the function execution, so it's safe)